health: #105 move "model" into "core" folder, add "blueprint" & "metadata" folders
