+++ src/core/public.ts
@@ -1,3 +1,4 @@
 export * from "./criteria/public";
+export * from "./entity/public";
 export * from "./expansion/public";
 export * from "./query/public";
