+++ test/facade/model/metadata.model.ts
@@ -1,4 +1,4 @@
-import { define } from "@entity-space/model";
+import { define } from "@entity-space/core";
 import { UserModel } from "./user.model";
 
 /**
