+++ test/facade/model/shape.model.ts
@@ -1,5 +1,5 @@
 import { Class } from "@entity-space/utils";
-import { define } from "@entity-space/model";
+import { define } from "@entity-space/core";
 import { DataEntryModel } from "./data-entry.model";
 import { AuthorModel } from "./author.model";
