+++ test/facade/model/tree-node-level.model.ts
@@ -1,4 +1,4 @@
-import { define } from "@entity-space/model";
+import { define } from "@entity-space/core";
 
 export class TreeNodeLevelModel {
     nodeId = define(Number, { id: true, required: true });
