+++ test/facade/model/user.model.ts
@@ -1,4 +1,4 @@
-import { define } from "@entity-space/model";
+import { define } from "@entity-space/core";
 
 export class UserModel {
     id = define(Number, { id: true, required: true, readOnly: true });
