+++ tsconfig-test.json
@@ -15,7 +15,7 @@
         "types": ["jasmine"],
         "lib": ["DOM", "scripthost", "ES2020"],
         "baseUrl": ".",
-        "paths": { "@entity-space/utils": ["src/utils/public.ts"], "@entity-space/core": ["src/core/public.ts"], "@entity-space/model": ["src/model/public.ts"] }
+        "paths": { "@entity-space/utils": ["src/utils/public.ts"], "@entity-space/core": ["src/core/public.ts"] }
     },
     "include": ["./src/**/*.ts", "./test/**/*.ts"]
 }
