+++ tsconfig.json
@@ -14,7 +14,7 @@
         "strict": true,
         "lib": ["DOM", "scripthost", "ES2020"],
         "baseUrl": ".",
-        "paths": { "@entity-space/utils": ["src/utils/public.ts"], "@entity-space/core": ["src/core/public.ts"], "@entity-space/model": ["src/model/public.ts"] }
+        "paths": { "@entity-space/utils": ["src/utils/public.ts"], "@entity-space/core": ["src/core/public.ts"] }
     },
     "include": ["./src/**/*.ts", "./test/**/*.ts"]
 }
